requirements.txt
python-telegram-bot==13.9
# pymongo==4.0.1
python-dotenv==0.19.0
flask==2.0.1
dash==1.21.0
stripe==2.61.0
pyTelegramBotAPI==4.7.1
werkzeug==0.16.1
